Save data to local storage
Tasks and task lists are now saved locally on the computer and
loaded after page reload and after closing and opening the page.
Bug was occuring when retrieving data and creating lists and tasks
out of old data because constructors for tasks and task lists were
calling function which saved data to local storage. Bug was successfully
resolved by sending an argument into constructors using which the
constructors know if they were called during the process of retrieving
data from the local storage. The saving to local storage function was
then not called if retrieving data was in process.